Refactoring RecordApp Source.
!We need know something of DuiLib then build app better.
