+++ demo/TestApp1/MenuWnd.cpp
@@ -1,6 +1,106 @@
 #include "stdafx.h"
+#include <tchar.h>
 #include "MenuWnd.h"
 #include "RecordWork.h"
+
+
+inline HBITMAP CreateMyBitmap(HDC hDC, int cx, int cy, COLORREF** pBits)
+{
+	LPBITMAPINFO lpbiSrc = NULL;
+	lpbiSrc = (LPBITMAPINFO) new BYTE[sizeof(BITMAPINFOHEADER)];
+	if (lpbiSrc == NULL) return NULL;
+
+	lpbiSrc->bmiHeader.biSize = sizeof(BITMAPINFOHEADER);
+	lpbiSrc->bmiHeader.biWidth = cx;
+	lpbiSrc->bmiHeader.biHeight = cy;
+	lpbiSrc->bmiHeader.biPlanes = 1;
+	lpbiSrc->bmiHeader.biBitCount = 32;
+	lpbiSrc->bmiHeader.biCompression = BI_RGB;
+	lpbiSrc->bmiHeader.biSizeImage = cx * cy;
+	lpbiSrc->bmiHeader.biXPelsPerMeter = 0;
+	lpbiSrc->bmiHeader.biYPelsPerMeter = 0;
+	lpbiSrc->bmiHeader.biClrUsed = 0;
+	lpbiSrc->bmiHeader.biClrImportant = 0;
+
+	HBITMAP hBitmap = CreateDIBSection(hDC, lpbiSrc, DIB_RGB_COLORS, (void **)pBits, NULL, NULL);
+	delete[] lpbiSrc;
+	return hBitmap;
+}
+
+CShadowWnd::CShadowWnd() 
+{
+	m_di.sDrawString = _T("file='menu_bk.png' corner='40,8,8,8'");
+}
+
+LPCTSTR CShadowWnd::GetWindowClassName() const 
+{ 
+	return _T("UIShadow");
+}
+
+UINT CShadowWnd::GetClassStyle() const 
+{
+	return UI_CLASSSTYLE_FRAME;
+}
+
+void CShadowWnd::OnFinalMessage(HWND /*hWnd*/) 
+{ 
+	delete this; 
+};
+
+void CShadowWnd::RePaint()
+{
+	RECT rcClient = { 0 };
+	::GetClientRect(m_hWnd, &rcClient);
+	DWORD dwWidth = rcClient.right - rcClient.left;
+	DWORD dwHeight = rcClient.bottom - rcClient.top;
+
+	HDC hDcPaint = ::GetDC(m_hWnd);
+	HDC hDcBackground = ::CreateCompatibleDC(hDcPaint);
+	COLORREF* pBackgroundBits;
+	HBITMAP hbmpBackground = CreateMyBitmap(hDcPaint, dwWidth, dwHeight, &pBackgroundBits);
+	::ZeroMemory(pBackgroundBits, dwWidth * dwHeight * 4);
+	HBITMAP hOldBitmap = (HBITMAP) ::SelectObject(hDcBackground, hbmpBackground);
+
+	CRenderEngine::DrawImage(hDcBackground, &m_pm, rcClient, rcClient, m_di);
+
+	RECT rcWnd = { 0 };
+	::GetWindowRect(m_hWnd, &rcWnd);
+
+	BLENDFUNCTION bf = { AC_SRC_OVER, 0, 255, AC_SRC_ALPHA };
+	POINT ptPos = { rcWnd.left, rcWnd.top };
+	SIZE sizeWnd = { dwWidth, dwHeight };
+	POINT ptSrc = { 0, 0 };
+	UpdateLayeredWindow(m_hWnd, hDcPaint, &ptPos, &sizeWnd, hDcBackground, &ptSrc, 0, &bf, ULW_ALPHA);
+
+	::SelectObject(hDcBackground, hOldBitmap);
+	if (hDcBackground != NULL) ::DeleteDC(hDcBackground);
+	if (hbmpBackground != NULL) ::DeleteObject(hbmpBackground);
+	::ReleaseDC(m_hWnd, hDcPaint);
+
+	m_bNeedUpdate = false;
+}
+
+LRESULT CShadowWnd::HandleMessage(UINT uMsg, WPARAM wParam, LPARAM lParam)
+{
+	if (uMsg == WM_CREATE) SetTimer(m_hWnd, 9000, 10, NULL);
+	else if (uMsg == WM_SIZE) m_bNeedUpdate = true;
+	else if (uMsg == WM_CLOSE) KillTimer(m_hWnd, 9000);
+	else if (uMsg == WM_TIMER) {
+		if (LOWORD(wParam) == 9000 && m_bNeedUpdate == true) {
+			if (!::IsIconic(m_hWnd)) RePaint();
+		}
+	}
+
+	return CWindowWnd::HandleMessage(uMsg, wParam, lParam);
+}
+
+
+CMenuWnd::CMenuWnd()
+:m_pOwner(nullptr),
+m_pShadowWnd(nullptr)
+{
+
+}
 
 void CMenuWnd::Init(CControlUI* pOwner, POINT pt) {
 	if (pOwner == NULL) return;
@@ -62,6 +162,7 @@ void CMenuWnd::OnFinalMessage(HWND /*hWnd*/)
 
 void CMenuWnd::Notify(TNotifyUI& msg)
 {
+	/*
 	if (msg.sType == _T("windowinit")) OnPrepare();
 	if (msg.sType == _T("click")) {
 		bool bClose = true;
@@ -177,6 +278,7 @@ void CMenuWnd::Notify(TNotifyUI& msg)
 			if (m_pOwner) m_pOwner->GetManager()->SendNotify(m_pOwner, _T("menu_Delete"), 0, 0, true);
 		}
 	}
+	*/
 }
 
 HWND CMenuWnd::Create(HWND hwndParent, LPCTSTR pstrName, DWORD dwStyle, DWORD dwExStyle, int x, int y, int cx, int cy, HMENU hMenu)
@@ -326,7 +428,7 @@ void CMenuWnd::SetChooseCode(LPCTSTR pDesLbName, LPCTSTR pLbName)
 
 void CMenuWnd::OnPrepare()
 {
-	if (NULL != m_Record->m_pCodeName)
+	/*if (NULL != m_Record->m_pCodeName)
 		ChooseCode(m_Record->m_pCodeName);
 	if (NULL != m_Record->m_pLzPageName)
 		ChooseLzPage(m_Record->m_pLzPageName);
@@ -351,5 +453,5 @@ void CMenuWnd::OnPrepare()
 	{
 		cLbUnMcf->SetBkImage(_T("record/choose.png"));
 		cLbMcf->SetBkImage(_T(""));
-	}
+	}*/
 }
